requirements.txt
app==0.0.1
argon2-cffi==20.1.0
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cffi==1.14.5
click==7.1.2
Cython==0.29.22
decorator==4.4.2
defusedxml==0.7.1
enum34==1.1.10
execnet==1.8.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
google-trans-new==1.1.9
grpcio==1.32.0
gunicorn
h5py==2.10.0
idna==2.10
imageio==2.9.0
imutils==0.5.4
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nibabel==3.2.1
notebook==6.3.0
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
path==15.1.2
path.py==12.5.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.1.1
plaidml==0.7.0
plaidml-keras==0.7.0
pluggy==0.13.1
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.7
psutil==5.8.0
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.1
Pygments==2.8.1
pylint==2.7.4
pylint-protobuf==0.20.0
pyrsistent==0.17.3
pyserial==3.5
pytest==6.2.2
pytest-shutil==1.7.0
python-dateutil==2.8.1
python-gdcm==3.0.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
smmap==4.0.0
sqlparse==0.4.1
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wxPython==4.1.0
